requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
codecov==2.1.8
colorama==0.4.3
coverage==5.2.1
cycler==0.10.0
fpl==0.6.22
idna==2.10
iniconfig==1.0.1
kiwisolver==1.2.0
matplotlib==3.3.1
more-itertools==8.4.0
multidict==4.7.6
numpy==1.19.1
packaging==20.4
pandas==1.1.1
Pillow==7.2.0
pluggy==0.13.1
PTable==0.9.2
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-aiohttp==0.3.0
pytest-cov==2.10.1
pytest-mock==3.2.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
scipy==1.5.2
seaborn==0.10.1
six==1.15.0
toml==0.10.1
urllib3==1.25.10
yarl==1.5.1
